requirements.txt
Django>=1.4.5,<1.9
djangorestframework>=2.3.8,<3.0
six>=1.10
django-filter>=0.11